webmidi/requestmidiaccess.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="UTF-8">
    <title>Web MIDI API Test: requestMIDIAccess()</title>
    <link rel="author" title="Ryoya KAWAI" href="mailto:ryoya.kawai@gmail.com">
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#requestmidiaccess"> <!-- 4.1 -->
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#midiaccess-interface"> <!-- 5 -->
    <meta name="assert" content="Return true when test success in requestMIDIAccess().">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <p>Success when requestMIDIAccess() exists and MIDIAccess Interface is correctly obtained .</p>
    <div id="log"></div>
    <!-- PAGE CONTENT -->
    <script type="text/javascript">
    var MIDI;
    var test_requestMIDIAccess = async_test("requestMIDIAccess({sysex: flase})");
    function testWMA() {
        var ret=navigator.requestMIDIAccess({sysex:false}).then( successCallback, errorCallback );
        // 4.1 requestMIDIAccess() (in { sysex:false })
        test(function(){
            assert_equals("Promise", ret.constructor.name);
        }, "Return type in {sysex:false}");
        function successCallback(access) {
            MIDI=access;
            // 5. MIDIAccess Interface
            test_requestMIDIAccess.step(function() {
                test(function(){
                    assert_equals("MIDIAccess", MIDI.constructor.name);
                }, "MIDIAccess Interface name test in {sysex:false}");
                // 5.1 Attributes
                test(function(){
                    assert_equals("MIDIInputMap", MIDI.inputs.constructor.name);
                    assert_equals("object", typeof MIDI.onconnect);
                    assert_equals("object", typeof MIDI.ondisconnect);
                    assert_equals("MIDIOutputMap", MIDI.outputs.constructor.name);
                    assert_false(MIDI.sysexEnabled);
                }, "MIDIAccess interface attribute test in {sysex:false}.");
            });
            test_requestMIDIAccess.done();
        }
        function errorCallback() {
            test_requestMIDIAccess.step(function() {
                assert_unreached();
                test_requestMIDIAccess.done();
            });
        }
    }
    testWMA();
    MIDI="";
    var test_requestMIDIAccess02 = async_test("requestMIDIAccess({sysex: true}) - W3C Spec: Web MIDI API");
    function testWMA02() {
        var ret02=navigator.requestMIDIAccess({sysex:true}).then( successCallback, errorCallback );
        // 4.1 requestMIDIAccess() (in { sysex:true })
        test(function(){
            assert_equals("Promise", ret02.constructor.name);
        }, "Return type test in sysex:true");
        function successCallback(access) {
            MIDI=access;
            // 5. MIDIAccess Interface
            test_requestMIDIAccess02.step(function() {
                test(function(){
                    assert_equals("MIDIAccess", MIDI.constructor.name);
                }, "MIDIAccess interface name test in {sysex:true}");
                // 5.1 Attributes
                test(function(){
                    assert_equals("MIDIInputMap", MIDI.inputs.constructor.name);
                    assert_equals("object", typeof MIDI.onconnect);
                    assert_equals("object", typeof MIDI.ondisconnect);
                    assert_equals("MIDIOutputMap", MIDI.outputs.constructor.name);
                    assert_true(MIDI.sysexEnabled);
                }, "MIDIAccess interface attribute test in {sysex:true}.");
                assert_true(true);
            });
            test_requestMIDIAccess02.done();
        }
        function errorCallback() {
            test_requestMIDIAccess02.step(function() {
                assert_false(true);
                test_requestMIDIAccess02.done();
            });
        }
    }
    testWMA02();
    </script>
  </body>
</html>
